+++ dart/geobase/lib/coordinates.dart
@@ -22,7 +22,6 @@ export 'src/codes/geo_representation.dart';
 export 'src/constants/epsilon.dart';
 export 'src/coordinates/base/aligned.dart';
 export 'src/coordinates/base/box.dart';
-export 'src/coordinates/base/measurable.dart';
 export 'src/coordinates/base/position.dart';
 export 'src/coordinates/base/positionable.dart';
 export 'src/coordinates/crs/coord_ref_sys.dart';